requirements.txt
# ================================================================================
# requirements.txt - Dependencias del Proyecto
# ================================================================================
# 
# Instalación:
#   pip install -r requirements.txt
#
# O instalación individual:
#   pip install flask requests beautifulsoup4 sentence-transformers groq numpy scikit-learn
#
# ================================================================================

# Framework Web
flask==3.0.0
Werkzeug==3.0.1

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Machine Learning & NLP
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.1.0

# IA Generativa
groq==0.36.0

# Procesamiento de Datos
numpy==1.24.3
scikit-learn==1.3.2

# Utilidades
urllib3==2.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6

# Opcional pero recomendado
python-dotenv==1.0.0  # Para manejar variables de entorno

# Telegram
python-telegram-bot==20.7